FB Homepage - server side
Today I'm just adding small part of FB homepage project implementing PHP
script showing the possibility of accessing received email and password
on server-side. Of course it’s possible to store it or exploit it
immediately. As for my version, this script is only sending it back to
user as proof and is not storing it in any way.